+++ app/src/main/AndroidManifest.xml
@@ -11,9 +11,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SelectOptionActivity"></activity>
+        <activity android:name=".CarAvailability"></activity>
         <activity android:name=".ScanQRActivity" />
-        <activity android:name=".CreateJobCard" />
+        <activity android:name=".Registration" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
