+++ app/src/main/java/com/example/vimaljagruti/MainActivity.java
@@ -10,7 +10,6 @@ import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
     CardView createJobCard,createBill,veiwJobCard,veiwBill;
@@ -30,7 +29,7 @@ public class MainActivity extends AppCompatActivity {
         createBill.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent=new Intent(getApplicationContext(),SelectOptionActivity.class);
+                Intent intent=new Intent(getApplicationContext(), CarAvailability.class);
                 startActivity(intent);
             }
         });
@@ -38,7 +37,7 @@ public class MainActivity extends AppCompatActivity {
         createJobCard.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent=new Intent(getApplicationContext(),SelectOptionActivity.class);
+                Intent intent=new Intent(getApplicationContext(), CarAvailability.class);
                 startActivity(intent);
             }
         });
