+++ app/src/main/res/layout/activity_registration.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/bgapps"
-    tools:context=".CreateJobCard">
+    tools:context=".Registration">
     <RelativeLayout
         android:layout_marginTop="60dp"
         android:layout_marginBottom="40dp"
